pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "imio"
dynamic = ["version"]
description = "A simple and robust image I/O and hashing utility library."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "Pillow",
]

[tool.setuptools.dynamic]
version = {attr = "imio.__version__"}